[Silver I] Title: 회의실 배정, Time: 28 ms, Memory: 3688 KB -BaekjoonHub

+++ 백준/Silver/1931. 회의실 배정/회의실 배정.cc
@@ -6,7 +6,7 @@ https://www.acmicpc.net/problem/1931
 
 using namespace std;
 
-using pll = pair<long long, long long>;
+using pll = pair<int, int>;
 
 bool comp(pll& p1, pll& p2)
 {
